# Improve performance by searching substrings instead of regex.