--- a/.travis.yml
+++ b/.travis.yml
@@ -4,3 +4,7 @@ node_js:
   - 4
   - 5
   - 6
+
+script:
+- npm run perf
+- npm test