--- a/package.json
+++ b/package.json
@@ -7,7 +7,8 @@
     "test": "test"
   },
   "scripts": {
-    "test": "node test"
+    "test": "node test",
+    "perf": "node perf"
   },
   "author": "Amjith Ramanujam",
   "license": "MIT",