@blainesch 2016-08-06
Also, given how fast this is, what are your thoughts on supporting a low n-gram type query?

I believe you are familiar, but when I type in `abcd` I'd do a search for `abcd`, `abc`, `bcd`, `ab`, `bc` and `cd`. Of course returning after the first match, preferring the larger match.