@amjith 2016-08-06
This is impressive. The code looks straight-forward and I like the search().for() semantic. 👍