@amjith 2016-08-06
I'm not opposed to the idea of having n-grams but I think it might start to screw with the ranking algorithm. 

I'd like it to be an optional feature to be turned on via options.